Revert "Added Backend"
This reverts commit 256707ba809d54c6cdc12f02dbf548a8198e0274.
