+++ app/(tabs)/education.tsx
