+++ app/(tabs)/news.tsx
