--- docker-compose.yml
@@ -1,54 +1,0 @@
-version: '3.8'
-
-services:
-  mongodb:
-    image: mongo:6.0
-    container_name: trading-simulator-db
-    restart: unless-stopped
-    ports:
-      - "27017:27017"
-    environment:
-      MONGO_INITDB_ROOT_USERNAME: admin
-      MONGO_INITDB_ROOT_PASSWORD: password123
-      MONGO_INITDB_DATABASE: trading_simulator
-    volumes:
-      - mongodb_data:/data/db
-      - ./database/init:/docker-entrypoint-initdb.d
-    networks:
-      - trading-network
-
-  backend:
-    build:
-      context: ./backend
-      dockerfile: Dockerfile
-    container_name: trading-simulator-backend
-    restart: unless-stopped
-    ports:
-      - "5000:5000"
-    environment:
-      - MONGO_URI=mongodb://admin:password123@mongodb:27017/trading_simulator?authSource=admin
-      - JWT_SECRET_KEY=your-super-secret-jwt-key-change-this-in-production
-      - FLASK_ENV=production
-    depends_on:
-      - mongodb
-    networks:
-      - trading-network
-
-  redis:
-    image: redis:7-alpine
-    container_name: trading-simulator-redis
-    restart: unless-stopped
-    ports:
-      - "6379:6379"
-    volumes:
-      - redis_data:/data
-    networks:
-      - trading-network
-
-volumes:
-  mongodb_data:
-  redis_data:
-
-networks:
-  trading-network:
-    driver: bridge
