--- frontend/contexts/AuthContext.tsx
@@ -1,105 +1,0 @@
-import React, { createContext, useContext, useEffect, useState, ReactNode } from 'react';
-import { authService, User } from '../services/authService';
-
-interface AuthContextType {
-  user: User | null;
-  isAuthenticated: boolean;
-  isLoading: boolean;
-  login: (email: string, password: string) => Promise<{ success: boolean; error?: string }>;
-  register: (name: string, email: string, password: string, confirmPassword: string) => Promise<{ success: boolean; error?: string }>;
-  logout: () => Promise<void>;
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-  return context;
-};
-
-interface AuthProviderProps {
-  children: ReactNode;
-}
-
-export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
-  const [user, setUser] = useState<User | null>(null);
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    checkAuthStatus();
-  }, []);
-
-  const checkAuthStatus = async () => {
-    try {
-      const isAuth = await authService.checkAuthStatus();
-      if (isAuth) {
-        const currentUser = authService.getCurrentUser();
-        setUser(currentUser);
-        setIsAuthenticated(true);
-      }
-    } catch (error) {
-      console.error('Auth check failed:', error);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const login = async (email: string, password: string) => {
-    try {
-      const result = await authService.login({ email, password });
-      if (result.success && result.user) {
-        setUser(result.user);
-        setIsAuthenticated(true);
-        return { success: true };
-      } else {
-        return { success: false, error: result.error || 'Login failed' };
-      }
-    } catch (error) {
-      return { success: false, error: 'Login failed. Please try again.' };
-    }
-  };
-
-  const register = async (name: string, email: string, password: string, confirmPassword: string) => {
-    try {
-      const result = await authService.register({ name, email, password, confirmPassword });
-      if (result.success && result.user) {
-        setUser(result.user);
-        setIsAuthenticated(true);
-        return { success: true };
-      } else {
-        return { success: false, error: result.error || 'Registration failed' };
-      }
-    } catch (error) {
-      return { success: false, error: 'Registration failed. Please try again.' };
-    }
-  };
-
-  const logout = async () => {
-    try {
-      await authService.logout();
-      setUser(null);
-      setIsAuthenticated(false);
-    } catch (error) {
-      console.error('Logout failed:', error);
-    }
-  };
-
-  const value: AuthContextType = {
-    user,
-    isAuthenticated,
-    isLoading,
-    login,
-    register,
-    logout,
-  };
-
-  return (
-    <AuthContext.Provider value={value}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
