+++ package.json
@@ -1,50 +1,49 @@
 {
-  "name": "trading-simulator-fullstack",
+  "name": "trading-simulator",
+  "main": "expo-router/entry",
   "version": "1.0.0",
-  "description": "A comprehensive trading simulator with React Native frontend, Flask backend, and MongoDB database",
-  "main": "index.js",
+  "private": true,
   "scripts": {
-    "dev": "concurrently \"npm run dev:backend\" \"npm run dev:frontend\"",
-    "dev:frontend": "cd frontend && npm run dev",
-    "dev:backend": "cd backend && python app.py",
-    "setup": "npm run setup:frontend && npm run setup:backend && npm run setup:database",
-    "setup:frontend": "cd frontend && npm install",
-    "setup:backend": "cd backend && pip install -r requirements.txt",
-    "setup:database": "cd database && python schemas.py",
-    "build": "npm run build:frontend",
-    "build:frontend": "cd frontend && npm run build:web",
-    "test": "npm run test:frontend && npm run test:backend",
-    "test:frontend": "cd frontend && npm test",
-    "test:backend": "cd backend && python -m pytest",
-    "lint": "npm run lint:frontend && npm run lint:backend",
-    "lint:frontend": "cd frontend && npm run lint",
-    "lint:backend": "cd backend && flake8 .",
-    "start": "npm run start:backend",
-    "start:backend": "cd backend && gunicorn app:app",
-    "start:frontend": "cd frontend && expo start"
+    "dev": "EXPO_NO_TELEMETRY=1 expo start",
+    "build:web": "expo export --platform web",
+    "lint": "expo lint"
   },
-  "keywords": [
-    "trading",
-    "simulator",
-    "options",
-    "react-native",
-    "flask",
-    "mongodb",
-    "finance",
-    "education",
-    "gamification"
-  ],
-  "author": "Your Name",
-  "license": "MIT",
+  "dependencies": {
+    "@expo/vector-icons": "^14.1.0",
+    "@lucide/lab": "^0.1.2",
+    "@react-navigation/bottom-tabs": "^7.2.0",
+    "@react-navigation/native": "^7.0.14",
+    "axios": "^1.12.1",
+    "expo": "^53.0.0",
+    "expo-blur": "~14.1.3",
+    "expo-camera": "~16.1.5",
+    "expo-constants": "~17.1.3",
+    "expo-font": "~13.2.2",
+    "expo-haptics": "~14.1.3",
+    "expo-linear-gradient": "~14.1.3",
+    "expo-linking": "~7.1.3",
+    "expo-router": "~5.0.2",
+    "expo-splash-screen": "~0.30.6",
+    "expo-status-bar": "~2.2.2",
+    "expo-symbols": "~0.4.3",
+    "expo-system-ui": "~5.0.5",
+    "expo-web-browser": "~14.1.5",
+    "lucide-react-native": "^0.475.0",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-native": "0.79.1",
+    "react-native-gesture-handler": "~2.24.0",
+    "react-native-reanimated": "~3.17.4",
+    "react-native-safe-area-context": "5.3.0",
+    "react-native-screens": "~4.10.0",
+    "react-native-svg": "15.11.2",
+    "react-native-url-polyfill": "^2.0.0",
+    "react-native-web": "^0.20.0",
+    "react-native-webview": "13.13.5"
+  },
   "devDependencies": {
-    "concurrently": "^8.2.2"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/your-username/trading-simulator.git"
-  },
-  "bugs": {
-    "url": "https://github.com/your-username/trading-simulator/issues"
-  },
-  "homepage": "https://github.com/your-username/trading-simulator#readme"
+    "@babel/core": "^7.25.2",
+    "@types/react": "~19.0.10",
+    "typescript": "~5.8.3"
+  }
 }
